Fixing, features, comments, cleaning
I was working with a big mistake that probably was the cause of "send(): bad file descriptor" error that i was dealing with. That is that in my user class i was using a pointer to the respective socket's element of a actual user that is stored in "_sockets" vector, to  access the "fd" field and the "revents" field. Now i realise that this was very stupid of me because vector container reallocates space so changing elements addresses. That lead to a memory access fault. I fixed it by implementing accessors in "Server" class.

I have fixed the redundant "address in use" error by setting the socket's options REUSEADDR and REUSEPORT with the setsockopt() function

I've completed the registration part of clients. But i was just focused on the big steps (architecture) of the server-to-client information part, not yet the deep treatment. It remains to complete the backend of that architecture by defining holders of the useful data information of the network.